% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{Ancestor}
\alias{Ancestor}
\title{Ancestor Object}
\usage{
Ancestor(resourceId = NULL)
}
\arguments{
\item{resourceId}{Resource id of the ancestor}
}
\value{
Ancestor object
}
\description{
Ancestor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifying information for a single ancestor of a project.
}
