% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{FolderOperation}
\alias{FolderOperation}
\title{FolderOperation Object}
\usage{
FolderOperation(operationType = NULL, displayName = NULL,
  sourceParent = NULL, destinationParent = NULL)
}
\arguments{
\item{operationType}{The type of this operation}

\item{displayName}{The display name of the folder}

\item{sourceParent}{The resource name of the folder's parent}

\item{destinationParent}{The resource name of the folder or organization we are either creating}
}
\value{
FolderOperation object
}
\description{
FolderOperation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata describing a long running folder operation
}
