% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{Lien}
\alias{Lien}
\title{Lien Object}
\usage{
Lien(parent = NULL, createTime = NULL, name = NULL, reason = NULL,
  origin = NULL, restrictions = NULL)
}
\arguments{
\item{parent}{A reference to the resource this Lien is attached to}

\item{createTime}{The creation time of this Lien}

\item{name}{A system-generated unique identifier for this Lien}

\item{reason}{Concise user-visible strings indicating why an action cannot be performed}

\item{origin}{A stable, user-visible/meaningful string identifying the origin of the}

\item{restrictions}{The types of operations which should be blocked as a result of this Lien}
}
\value{
Lien object
}
\description{
Lien Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Lien represents an encumbrance on the actions that can be performed on aresource.
}
\seealso{
Other Lien functions: \code{\link{liens.create}}
}
