% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_functions.R
\name{liens.create}
\alias{liens.create}
\title{Create a Lien which applies to the resource denoted by the `parent` field.Callers of this method will require permission on the `parent` resource.For example, applying to `projects/1234` requires permission`resourcemanager.projects.updateLiens`.NOTE: Some resources may limit the number of Liens which may be applied.}
\usage{
liens.create(Lien)
}
\arguments{
\item{Lien}{The \link{Lien} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/resource-manager}{Google Documentation}

Other Lien functions: \code{\link{Lien}}
}
