% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_functions.R
\name{liens.list}
\alias{liens.list}
\title{List all Liens applied to the `parent` resource.Callers of this method will require permission on the `parent` resource.For example, a Lien with a `parent` of `projects/1234` requires permission`resourcemanager.projects.get`.}
\usage{
liens.list(parent = NULL, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{parent}{The name of the resource to list all attached Liens}

\item{pageToken}{The `next_page_token` value returned from a previous List request, if any}

\item{pageSize}{The maximum number of items to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/resource-manager}{Google Documentation}
}
