% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{ListLiensResponse}
\alias{ListLiensResponse}
\title{ListLiensResponse Object}
\usage{
ListLiensResponse(nextPageToken = NULL, liens = NULL)
}
\arguments{
\item{nextPageToken}{Token to retrieve the next page of results, or empty if there are no more}

\item{liens}{A list of Liens}
}
\value{
ListLiensResponse object
}
\description{
ListLiensResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for Liens.ListLiens.
}
