% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{ListProjectsResponse}
\alias{ListProjectsResponse}
\title{ListProjectsResponse Object}
\usage{
ListProjectsResponse(nextPageToken = NULL, projects = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token}

\item{projects}{The list of Projects that matched the list filter}
}
\value{
ListProjectsResponse object
}
\description{
ListProjectsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A page of the response received from theListProjectsmethod.A paginated response where more pages are available has`next_page_token` set. This token can be used in a subsequent request toretrieve the next request page.
}
