% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{Operation}
\alias{Operation}
\title{Google Cloud Resource Manager API Objects 
The Google Cloud Resource Manager API provides methods for creating, reading, and updating project metadata.}
\usage{
Operation(Operation.response = NULL, Operation.metadata = NULL,
  response = NULL, name = NULL, error = NULL, metadata = NULL,
  done = NULL)
}
\arguments{
\item{Operation.response}{The \link{Operation.response} object or list of objects}

\item{Operation.metadata}{The \link{Operation.metadata} object or list of objects}

\item{response}{The normal response of the operation in case of success}

\item{name}{The server-assigned name, which is only unique within the same service that}

\item{error}{The error result of the operation in case of failure or cancellation}

\item{metadata}{Service-specific metadata associated with the operation}

\item{done}{If the value is `false`, it means the operation is still in progress}
}
\value{
Operation object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:33:16
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudresourcemanagerv1.auto/R/cloudresourcemanager_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Operation Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This resource represents a long-running operation that is the result of anetwork API call.
}
\seealso{
Other Operation functions: \code{\link{Operation.metadata}},
  \code{\link{Operation.response}}
}
