% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{Organization}
\alias{Organization}
\title{Organization Object}
\usage{
Organization(owner = NULL, lifecycleState = NULL, name = NULL,
  displayName = NULL, creationTime = NULL)
}
\arguments{
\item{owner}{The owner of this Organization}

\item{lifecycleState}{The organization's current lifecycle state}

\item{name}{Output Only}

\item{displayName}{A friendly string to be used to refer to the Organization in the UI}

\item{creationTime}{Timestamp when the Organization was created}
}
\value{
Organization object
}
\description{
Organization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The root node in the resource hierarchy to which a particular entity's(e.g., company) resources belong.
}
