% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{OrganizationOwner}
\alias{OrganizationOwner}
\title{OrganizationOwner Object}
\usage{
OrganizationOwner(directoryCustomerId = NULL)
}
\arguments{
\item{directoryCustomerId}{The Google for Work customer id used in the Directory API}
}
\value{
OrganizationOwner object
}
\description{
OrganizationOwner Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The entity that owns an Organization. The lifetime of the Organization andall of its descendants are bound to the `OrganizationOwner`. If the`OrganizationOwner` is deleted, the Organization and all its descendants willbe deleted.
}
