% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{Project}
\alias{Project}
\title{Google Cloud Resource Manager API Objects 
The Google Cloud Resource Manager API provides methods for creating, reading, and updating project metadata.}
\usage{
Project(Project.labels = NULL, projectNumber = NULL, projectId = NULL,
  lifecycleState = NULL, name = NULL, createTime = NULL, labels = NULL,
  parent = NULL)
}
\arguments{
\item{Project.labels}{The \link{Project.labels} object or list of objects}

\item{projectNumber}{The number uniquely identifying the project}

\item{projectId}{The unique, user-assigned ID of the Project}

\item{lifecycleState}{The Project lifecycle state}

\item{name}{The user-assigned display name of the Project}

\item{createTime}{Creation time}

\item{labels}{The labels associated with this Project}

\item{parent}{An optional reference to a parent Resource}
}
\value{
Project object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:17:15
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudresourcemanagerv1.auto/R/cloudresourcemanager_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Project Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Project is a high-level Google Cloud Platform entity. It is a container for ACLs, APIs, AppEngine Apps, VMs, and other Google Cloud Platform resources.
}
\seealso{
Other Project functions: \code{\link{Project.labels}},
  \code{\link{projects.update}}
}

