% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{ProjectCreationStatus}
\alias{ProjectCreationStatus}
\title{ProjectCreationStatus Object}
\usage{
ProjectCreationStatus(ready = NULL, createTime = NULL, gettable = NULL)
}
\arguments{
\item{ready}{True if the project creation process is complete}

\item{createTime}{Creation time of the project creation workflow}

\item{gettable}{True if the project can be retrieved using GetProject}
}
\value{
ProjectCreationStatus object
}
\description{
ProjectCreationStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A status object which is used as the `metadata` field for the Operationreturned by CreateProject. It provides insight for when significant phases ofProject creation have completed.
}
