% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_functions.R
\name{projects.create}
\alias{projects.create}
\title{Request that a new Project be created. The result is an Operation whichcan be used to track the creation process. It is automatically deletedafter a few hours, so there is no need to call DeleteOperation.Our SLO permits Project creation to take up to 30 seconds at the 90thpercentile. As of 2016-08-29, we are observing 6 seconds 50th percentilelatency. 95th percentile latency is around 11 seconds. We recommendpolling at the 5th second with an exponential backoff.}
\usage{
projects.create(Project)
}
\arguments{
\item{Project}{The \link{Project} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/resource-manager}{Google Documentation}

Other Project functions: \code{\link{Project.labels}},
  \code{\link{Project}}, \code{\link{projects.update}}
}
