% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_functions.R
\name{projects.getAncestry}
\alias{projects.getAncestry}
\title{Gets a list of ancestors in the resource hierarchy for the Projectidentified by the specified `project_id` (for example, `my-project-123`).The caller must have read permissions for this Project.}
\usage{
projects.getAncestry(GetAncestryRequest, projectId)
}
\arguments{
\item{GetAncestryRequest}{The \link{GetAncestryRequest} object to pass to this method}

\item{projectId}{The Project ID (for example, `my-project-123`)}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/resource-manager}{Google Documentation}

Other GetAncestryRequest functions: \code{\link{GetAncestryRequest}}
}
