% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{SearchOrganizationsRequest}
\alias{SearchOrganizationsRequest}
\title{SearchOrganizationsRequest Object}
\usage{
SearchOrganizationsRequest(pageToken = NULL, pageSize = NULL,
  filter = NULL)
}
\arguments{
\item{pageToken}{A pagination token returned from a previous call to `SearchOrganizations`}

\item{pageSize}{The maximum number of Organizations to return in the response}

\item{filter}{An optional query string used to filter the Organizations to return in}
}
\value{
SearchOrganizationsRequest object
}
\description{
SearchOrganizationsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request sent to the `SearchOrganizations` method.
}
\seealso{
Other SearchOrganizationsRequest functions: \code{\link{organizations.search}}
}
