% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{SearchOrganizationsResponse}
\alias{SearchOrganizationsResponse}
\title{SearchOrganizationsResponse Object}
\usage{
SearchOrganizationsResponse(nextPageToken = NULL, organizations = NULL)
}
\arguments{
\item{nextPageToken}{A pagination token to be used to retrieve the next page of results}

\item{organizations}{The list of Organizations that matched the search query, possibly}
}
\value{
SearchOrganizationsResponse object
}
\description{
SearchOrganizationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response returned from the `SearchOrganizations` method.
}
